move get_cli_string out of cli.py
