+++ dotenv/cli.py
@@ -1,6 +1,10 @@
 import os
 
-import click
+try:
+    import click
+except ImportError:
+    raise ImportError('It seems python-dotenv is not installed with cli option. \
+                       Run pip install "python-dotenv[cli]" to fix this.')
 
 from .main import dotenv_values, get_key, set_key, unset_key
 
@@ -70,30 +74,6 @@ def unset(ctx, key):
         click.echo("Successfully removed %s" % key)
     else:
         exit(1)
-
-
-def get_cli_string(path=None, action=None, key=None, value=None, quote=None):
-    """Returns a string suitable for running as a shell script.
-
-    Useful for converting a arguments passed to a fabric task
-    to be passed to a `local` or `run` command.
-    """
-    command = ['dotenv']
-    if quote:
-        command.append('-q %s' % quote)
-    if path:
-        command.append('-f %s' % path)
-    if action:
-        command.append(action)
-        if key:
-            command.append(key)
-            if value:
-                if ' ' in value:
-                    command.append('"%s"' % value)
-                else:
-                    command.append(value)
-
-    return ' '.join(command).strip()
 
 
 if __name__ == "__main__":
